+++ src/components/ExampleHero.tsx
